+++ resources/js/components/ModalActionResponse.vue
@@ -1,51 +1,55 @@
 <template>
-    <Modal @modal-close="handleClose"
-           tabindex="-1"
-           role="dialog"
-           :size="data.size ?? '2xl'">
-        <div
-            @submit.prevent="handleConfirm"
-            slot-scope="props"
-            class="bg-white rounded-lg shadow-lg overflow-hidden dark:bg-gray-800 space-y-6"
-        >
-            <ModalHeader v-text="data.title"/>
-            <ModalContent>
-                <p v-if="data.html" v-html="data.html" />
-                <div v-if="data.body" v-text="data.body" />
+    <Modal :show="show"
+           :size="data.size ?? '2xl'"
+           @close-via-escape="handleClose"
+           role="alertdialog">
+        <div class="mx-auto bg-white dark:bg-gray-800 rounded-lg shadow-lg overflow-hidden">
+            <slot>
+                <ModalHeader v-text="data.title" />
                 <pre v-if="data.code">
                     <code v-text="data.code"></code>
                 </pre>
-            </ModalContent>
-            <ModalFooter>
-                <div class="flex items-center ml-auto">
-                    <CancelButton
-                        component="button"
-                        type="button"
-                        dusk="cancel-action-button"
-                        class="ml-auto mr-3"
-                        @click="$emit('close')"
-                    >
-                        Cancel
-                    </CancelButton>
+                <div v-else class="py-3 px-8">
+                    <div v-if="data.html" v-html="data.html"/>
+                    <p v-if="data.body" v-text="data.body"/>
+                </div>
+            </slot>
 
+            <ModalFooter>
+                <div class="ml-auto">
+                    <Button
+                        variant="link"
+                        state="mellow"
+                        @click.prevent="handleClose"
+                        class="mr-3"
+                        dusk="close-modal-button"
+                    >
+                        {{ data.closeButtonText ?? 'Close' }}
+                    </Button>
                 </div>
             </ModalFooter>
         </div>
-
     </Modal>
 </template>
 
 <script>
-export default {
-    name: "ModalActionResponse",
+import { Button } from 'laravel-nova-ui'
 
-    props: ["data"],
+export default {
+    components: {
+        Button,
+    },
+
+    emits: ['confirm', 'close'],
+
+    props: {
+        show: { type: Boolean, default: false },
+        data: { type: Object, required: true }
+    },
+
     methods: {
         handleClose() {
             this.$emit('close')
-        },
-        handleConfirm() {
-            this.$emit('confirm')
         },
     },
 }
@@ -56,7 +60,6 @@ pre {
     color: #f8fafc;
     background-color: #1e293b;
     overflow-x: auto;
-    border-radius: 0.75rem;
     padding: 1.25rem;
     box-shadow: 0 4px 6px -1px rgb(0 0 0 / 10%), 0 2px 4px -2px rgb(0 0 0 / 10%);
     display: flex;