+++ webpack.mix.js
@@ -1,7 +1,8 @@
 let mix = require('laravel-mix')
 let path = require('path')
+let NovaExtension = require('laravel-nova-devtool')
 
-require('./nova.mix')
+mix.extend('nova', new NovaExtension())
 
 mix
   .setPublicPath('dist')