+++ CHANGELOG.md
@@ -19,7 +19,7 @@
 ### Added
 - Added support for `laravel/nova:^5.0`.
 
-### Removed 
+### Removed
 - Removed support for Laravel 8 & 9
 
 ## [v0.3.0 (2024-03-13)](https://github.com/markwalet/nova-modal-response/compare/v0.2.2...v0.3.0)