+++ CHANGELOG.md
@@ -3,16 +3,19 @@
 ## [Unreleased](https://github.com/markwalet/nova-modal-response/compare/v0.3.0...master)
 
 ### Added
-- Installed Nova devtools
 - Close modal on escape
 - Added option to configure close button text
-- Added optional syntax highlighting for code
+- Added optional syntax highlighting for code (enabled by default)
 - Added dark mode support
+- Installed Nova devtools
 
 ### Changed
 - Moved to new webpack mix configuration
 - Made code block the full width of a modal
 - Changed default close button text to `Close` instead of `Cancel`
+
+### Removed
+- Removed support for `laravel/nova` V4
 
 ## [v0.3.1 (2024-12-16)](https://github.com/markwalet/nova-modal-response/compare/v0.3.0...v0.3.1)