+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## [Unreleased](https://github.com/markwalet/nova-modal-response/compare/v0.3.0...master)
 
+## [v1.0.0 (2024-12-20)](https://github.com/markwalet/nova-modal-response/compare/v0.3.0...v1.0.1)
+
 ### Added
 - Close modal on escape
 - Added option to configure close button text